refactor(daemon): extract standing-context assembly into a pure module
handle() built the per-turn standing context inline: the memory-index
injection template, the agent meta block, the collaboration guidance, the
parent-reply obligation, and the final resume/session/meta assembly — ~210
lines of string building interleaved with the session lifecycle.

Move all of it to session/turn/standing-context.ts behind a pure
buildStandingContext(input). The input is an explicit plain object; handle()
keeps only the impure lookups (store.getDisplayNames, slackBotUserIdFor,
runtimeEnvFor secret planning, host.usesMetaSystemPrompt) and computes them
before the call. encodeMemoryBoundaryBody and AGENT_META_OPENING move with
the text they serve; isStandingContextTitleEcho re-imports the latter.

No behavior change — the produced strings are byte-identical.

Verified: daemon typecheck clean; session-manager (95), no-response (3),
thread-context (5), plus daemon-smoke / mcp-tools / acp-host /
evaluation-runner (117) all pass; prettier + eslint clean on both files.
